Move deciles utils from web to data
Now that these utils don't derive information about visual encodings
(e.g. stroke colour, stroke dash), it makes sense to move them from web
to data; they contain business logic, and business logic lives in data.
